+++ lib/deprecate_soft.rb
@@ -1,24 +1,11 @@
 # frozen_string_literal: true
 
-require_relative 'deprecate_soft/version'
-
-require_relative 'deprecate_soft/method_wrapper'
+# require all files under lib/deprecate_soft
+Dir[File.join(__dir__, 'deprecate_soft', '*.rb')].sort.each do |file|
+  require file
+end
 
 module DeprecateSoft
-  def configure_base(base)
-    base.extend(ClassMethods)
-    base.extend(InstanceMethods)
-  end
-  module_function :configure_base
-
-  def included(base)
-    configure_base(base)
-  end
-
-  def extended(base)
-    configure_base(base)
-  end
-
   class << self
     attr_accessor :before_hook, :after_hook
     attr_writer :prefix, :suffix
@@ -34,23 +21,58 @@ module DeprecateSoft
     def configure
       yield self
     end
-  end
 
-  module InstanceMethods
-    def deprecate_soft(method_name, message)
-      #  protect against declaring deprecate_soft before method is defined
-      return unless method_defined?(method_name) || private_method_defined?(method_name)
+    def prefixed_name(method_name)
+      "#{prefix}#{method_name}_#{suffix}".to_sym
+    end
 
-      DeprecateSoft::MethodWrapper.wrap_method(self, method_name, message, is_class_method: false)
+    def included(base)
+      base.extend(ClassMethods)
+
+      base.define_singleton_method(:method_added) do |method_name|
+        pending = base.instance_variable_get(:@__pending_soft_wraps)
+        if pending&.key?(method_name)
+          DeprecateSoft::MethodWrapper.wrap_method(base, method_name, pending.delete(method_name), is_class_method: false)
+        end
+        super(method_name) if defined?(super)
+      end
+
+      base.singleton_class.define_method(:singleton_method_added) do |method_name|
+        pending = instance_variable_get(:@_pending_soft_wraps)
+        if pending&.key?(method_name)
+          DeprecateSoft::MethodWrapper.wrap_method(base, method_name, pending.delete(method_name), is_class_method: true)
+        end
+        super(method_name) if defined?(super)
+      end
     end
   end
 
   module ClassMethods
-    def deprecate_soft(method_name, message)
-      #  protect against declaring deprecate_soft before method is defined
-      return unless singleton_class.method_defined?(method_name) || singleton_class.private_method_defined?(method_name)
+    def deprecate_soft(method_name, message = nil)
+      hidden = DeprecateSoft.prefixed_name(method_name)
 
-      DeprecateSoft::MethodWrapper.wrap_method(self, method_name, message, is_class_method: true)
+      if method_defined?(method_name) || private_method_defined?(method_name)
+        return if method_defined?(hidden) || private_method_defined?(hidden)
+
+        DeprecateSoft::MethodWrapper.wrap_method(self, method_name, message, is_class_method: false)
+      else
+        @__pending_soft_wraps ||= {}
+        @__pending_soft_wraps[method_name] = message
+      end
+    end
+
+    def deprecate_class_soft(method_name, message = nil)
+      hidden = DeprecateSoft.prefixed_name(method_name)
+      target = singleton_class
+
+      if target.method_defined?(method_name) || target.private_method_defined?(method_name)
+        return if target.method_defined?(hidden) || target.private_method_defined?(hidden)
+
+        DeprecateSoft::MethodWrapper.wrap_method(self, method_name, message, is_class_method: true)
+      else
+        @_pending_soft_wraps ||= {}
+        @_pending_soft_wraps[method_name] = message
+      end
     end
   end
 end