+++ spec/redis_spec.rb
@@ -7,13 +7,27 @@ require 'redis'
 
 RSpec.describe 'DeprecateSoft with Redis tracking' do
   let(:mock_redis) { instance_double(Redis) }
+  let(:klass) do
+    Class.new do
+      include DeprecateSoft
+
+      def foo(a)
+        "x#{a}"
+      end
+
+      def self.name
+        'TestKlass' # for string formatting in the wrapper
+      end
+
+      deprecate_soft :foo, 'will be deleted'
+    end
+  end
 
   before do
     allow(mock_redis).to receive(:incr)
 
     DeprecateSoft.configure do |config|
       config.before_hook = lambda do |method, _message, args:|
-        # Replace # with : for Redis
         redis_key = "deprecate_soft:#{method}".gsub('#', ':')
         mock_redis.incr("#{redis_key}:#{Date.today}")
       end
@@ -22,22 +36,10 @@ RSpec.describe 'DeprecateSoft with Redis tracking' do
     end
   end
 
-  before do
-    stub_const('Klass', Class.new do
-      include DeprecateSoft
-
-      def foo(a)
-        "x#{a}"
-      end
-
-      deprecate_soft :foo, 'migrate to #future'
-    end)
-  end
-
   it 'increments the redis key on call' do
-    Klass.new.foo('42')
+    klass.new.foo('42')
 
-    expected_key = "deprecate_soft:Klass:foo:#{Date.today}"
+    expected_key = "deprecate_soft:TestKlass.foo:#{Date.today}"
     expect(mock_redis).to have_received(:incr).with(expected_key)
   end
 end